generateBodies functions filled down to component interface level. Anchor body is being generated and objectId is being passed down from physics engine for collisino filtering

+++ pkg_Simulator/include/sdsmt_simulator/property.h
@@ -9,7 +9,7 @@
 
 #include "model.h"
 
-typedef uint64_t object_id;
+typedef uint32_t object_id;
 
 class PropertyInfo
 {
